refactor: address review feedback
- Move onboardingDiscoverPosts mutation into users schema instead of a
  dedicated onboarding module
- Encrypt RECSWIPE_ORIGIN in Pulumi config to match the convention used
  for other internal service URLs
